requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
